requirements.txt
spotipy
requests
Flask
pyjwt